added loop to retry fetching audit logs
getExecutorForRoleChangeFromAuditLog() would attempt to locate the matching audit log entry immediately when called. Discord API can be delayed in the creation of the audit log entry (or it may never even happen). When this function is relied upon to ensure a user is tanked, this has already resulted in a record not being created for a tanked user, and no message occurs either. Severely delayed audit logs can cause this to occur more frequently.

This new fix forces the function to run Guild.fetchAuditLogs() up to 5 times before it considers the audit log entry non-existent, and waits 1000ms between each iteration. Also, moved the bypassingGMU check to occur only if the audit log entry is the matching one we were wanting, otherwise byBot could be set from a previous entry that is similar but not the right one, and any failure to find the audit log entry would result in being treated as a bypassing user and therefore never get logged or handled.

This update requires callers to check for rejection result "BYPASS" or "FAILURE" and handle accordingly.